Fix compiler warnings and a couple other minor clean-ups and improvements.

+++ addon/engine/src/Random.cpp
@@ -7,7 +7,7 @@
 int Random::Range(int min, int max)
 {
     assert(min <= max);
-    static std::default_random_engine rng(std::chrono::system_clock::now().time_since_epoch().count());
+    static std::default_random_engine rng((unsigned int)std::chrono::system_clock::now().time_since_epoch().count());
     std::uniform_int_distribution<int> distribution(min, max);
     return distribution(rng);
 }
@@ -15,7 +15,7 @@ int Random::Range(int min, int max)
 unsigned int Random::Range(unsigned int min, unsigned int max)
 {
     assert(min <= max);
-    static std::default_random_engine rng(std::chrono::system_clock::now().time_since_epoch().count());
+    static std::default_random_engine rng((unsigned int)std::chrono::system_clock::now().time_since_epoch().count());
     std::uniform_int_distribution<unsigned int> distribution(min, max);
     return distribution(rng);
 }
@@ -23,7 +23,7 @@ unsigned int Random::Range(unsigned int min, unsigned int max)
 float Random::Range(float min, float max)
 {
     assert(min <= max);
-    static std::default_random_engine rng(std::chrono::system_clock::now().time_since_epoch().count());
+    static std::default_random_engine rng((unsigned int)std::chrono::system_clock::now().time_since_epoch().count());
     std::uniform_real_distribution<float> distribution(min, max);
     return distribution(rng);
 }
